+++ test/post/test_post_open_message.js
@@ -79,8 +79,8 @@ const tests = [
     description: 'Post channel open message to Telegram',
     expected: {
       text: [
-        '🌹 node1',
-        `Accepted new 0.00000001 private channel from alias ${pubKey}. Inbound liquidity now: 0.00000001. Outbound liquidity now: 0.00000001.`,
+        `🌹 Accepted new 0.00000001 private channel from alias ${pubKey}. Inbound liquidity now: 0.00000001. Outbound liquidity now: 0.00000001.`,
+        'node1',
       ],
     },
   },
@@ -89,8 +89,8 @@ const tests = [
     description: 'Post channel open message to Telegram for self-channel',
     expected: {
       text: [
-        '🌹 node1',
-        `Opened new 0.00000001 channel to alias ${pubKey}. Inbound liquidity now: 0.00000001. Outbound liquidity now: 0.00000001.`,
+        `🌹 Opened new 0.00000001 channel to alias ${pubKey}. Inbound liquidity now: 0.00000001. Outbound liquidity now: 0.00000001.`,
+        'node1',
       ],
     },
   },